requirements.txt
streamlit==0.62.0
pillow==6.2.0
pytesseract==0.3.6